backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
pydantic
python-jose
passlib[bcrypt]
python-multipart
tensorflow-cpu
pymongo

